Fix - Lambda, DB, ComputeCluster bug fixes. (#163)
* moving into latest structure

* Delete resource.ftl artifact

This file has been moved, and yet remains in github in its old location. Local git refuses to download it so manually removing it.

* Update resource.ftl

File is missing changes from a file restructure. Manually re-applying them.

* refactor: allow for secrets with no prefix

* refactor: create parameter lookups for all fragment settings that are vault secrets

* fix: list keys no longer needs to perform storageId reference lookup

* feat(db): add db fqdn output

* refactor(db): username to incl. fqdn

db username attribute to be in user@host format

* fix(computecluster): use id reference

format a connection string with the full resource ID, not the reference.