Fix - ComputeCluster Env Variables (#170)
* moving into latest structure

* Delete resource.ftl artifact

This file has been moved, and yet remains in github in its old location. Local git refuses to download it so manually removing it.

* Update resource.ftl

File is missing changes from a file restructure. Manually re-applying them.

* Fix pseudo outputs (#167)

* fix(outputs): arm outputs to exclude pseudo resources
armOutput to only be called on non-pseudo resources.

* Fix - DB FQDN (#168)

* fix(db): fix formatting on retrieval of fqdn

* fix(computecluster): export env vars and pass to env file

export env  vars in shell and pass to environment file as source cmd not
avail in /bin/sh

Co-authored-by: Ross Murray <rossmurray@Rosss-MacBook-Pro.local>
Co-authored-by: Ross Murray <rossmurray@protonmail.com>