feat(tests): add loading of mocked utility functions
Some shell script utility functions are invoked during
prologue generation passes. Functions that perform tasks
on cloud services cannot be performed during a test.

This commit adds a mocked version of the invoked
function and sources this file. The function will create
the file expected, and allow the tests to continue.