fix(outputs): fix retrieval of pseudo resource outputs
Some resources are defined with a "pseudo" resource profile. This allows
them to use all the capabilities of resource profiles as defined in this
plugin, whilst they are not technically a resource as far as Azure
Services are concerned.

Presently, the "getReference" function does not allow for the fact that
an engine pass might attempt to retrieve a pseudo resource output. In
the case that it does (as it presently does on first
generating/deploying the baseline, which has a pseudo resource for the
"seed") it will incorrectly construct an ARM function, rather than
returning the results of the stack-outputs.

This commit updates getReference() with a check to see if its working
with a pseudo resource profile. If it is, it calls getExistingReference
which will correctly either:
  a) return empty; or
  b) return the discoverd value.