fix(baseline): segment seed either does/doesn't exist
The segment seed is a single value unique to each segment that allows
the engine to ensure global uniqueness of resource names as necessary
(amongst other uses). On the baseline component in this provider, the
seed value becomes a part of the globally unique resources such as the
KeyVault.

In #202 the azure-specific implementation of the getReference() function
was refactored to be easier to use. During this update, the call made
here was updated from "getExistingReference()" to use "getReference()".
This was incorrect and it should have remained as-is.

"getReference()" can:
a) construct an ARM-function reference to a resourceThe segment seed is
a single value unique to each segment that allows the engine to ensure
global uniqueness of resource names as necessary (amongst other uses).
On the baseline component in this provider, the seed value becomes a
part of the globally unique resources such as the KeyVault.

In #202 the azure-specific implementation of the getReference() function
was refactored to be easier to use. During this update, the call made
here was updated from "getExistingReference()" to use "getReference()".
This was incorrect and it should have remained as-is.

"getReference()" can:
a) construct an ARM-function reference to a resourceThe segment seed is
a single value unique to each segment that allows the engine to ensure
global uniqueness of resource names as necessary (amongst other uses).
On the baseline component in this provider, the seed value becomes a
part of the globally unique resources such as the KeyVault.

In #202 the azure-specific implementation of the getReference() function
was refactored to be easier to use. During this update, the call made
here was updated from "getExistingReference()" to use "getReference()".
This was incorrect and it should have remained as-is.

"getReference()" is used where the desired output might already exist
(it actually calls getExistingReference for this) but if nothing is
found, it will format an ARM function that can be used to reference the
target within an ARM template.

"getExistingReference" on the other hand is "Go and see if this has
already been deployed". It either has been and there is a stack-output
for the Id, or there isn't and an empty value is returned. This is what
we want in the event of the seed - it should never be capable of
returning an ARM function, and this is the same for all pseudo resource
outputs.