fix(tests): mocked output ids should be in template scope
In order to determine how specific an ARM resource reference must be,
the scope between the source resource is compared with the target
resource. Any differences of the subscription, the resource group or
region determine a different scope. The difference is then summarised as
a "level". A level of "template" is default and means that the target
and source are in the same template, whilst anything above that means
the level is the highest difference - i.e a level of "subscription"
indicates that the target resource is in a different subscription from
the source.

This becomes a problem when providing mocked values for the stack
outputs. Prior to this commit, the mocked value was a hard-coded and
nonexistent subscription and resourceGroup. When the scopes are compared
with the current deploymentUnit, this will always result in the highest
possible level of difference - subscription.

Subscription-level resources are implemented via "nesting" ARM templates
within eachother. There is an ARM Resource for a "template" which allows
for easily nesting templates whilst handling them as if they are just
resources. However this also requires recursively invoking the macro
@armResource, to pass the nested information into a ARM Resource for the
parent template.

During this recursive call, anything on the resource that causes the
scopes to be evaluated again will yet-again set the highest level of
scope difference ("subscription") and the cycle will begin again
indefinately.

This commit will align the mock resource Id value's parts (specifically
the subscription and the resourceGroup) with the current deployment
unit and the azure provider's mock inputsource configuration. When in
alignment, they will not indefinately recurse.