fix(tests): standardise mock values
Prior to this commit, mocked values were hard-coded in numerous places
but were required to align to ensure the tests performed as expected.
Mocked values are apart of other strings, so changing one meant having
to update the other. This is quite a fragile practice.

This commit standardises most the values used currently across the azure
testing provider by setting global variables in a single location and
using them throughout. Now there is a central place to update in the
future and a less fragile pattern to continue to use and develop with
future tests.