fix: move mock input types into azure provider
To ensure consistency for data types when using mocked input sources,
the Azure provider defines a number of global variables formatted
appropriately for that type.

These types are currently defined mostly in the azuretest provider,
as testing is the most-common use-case for using the mock input
source.

This poses a problem for new entrances that may make use of the
mock inputsource, as they will require both azure and azuretest
providers to be loaded.

This change moves the definitions for Mock types into the azure
provider, so they can be used by other entrances for non-testing
capabilities. The result is that in order to use mock, you only
have to load the azure provider, but to perform tests, you must
load both (already a requirement).