fix: establish resourcegroup name in the engine
Previously the Azure Resource Manager ResourceGroup - a name given to
the collection of resources that are being deployed together and a
"collection" heading under which they will be grouped - was generated
within the Executor.

As we have all the necessary information within the engine at runtime
and have little reason to override this value, we should instead define
it within the engine, making it available everywhere else.

This was previously an issue for Pseudo Stacks that would define the
ResourceGroup as an empty string should any of the various context-based
scripts fail. Whilst that is an issue of its own (resolved separately)
moving this logic to the engine will remove unnecessary complexity for
the executor.