refactor: disable purge protection on keyvault
Purge Protection on keyvault is to prevent a secret / key from being
irrovocably removed. However, once a secret / key has been deleted (from
say a "stop" deploy mode) there are now only two options to move
forward:

1) Purge (now unavailable)
2) Recover the secret

Typically if a "stop" mode deployment is performed you want to at some
stage run a "create" or "update", neither of which should have to check
for deleted but unpurged secrets with the same name before creating
them.

This disabled purge protection on the baseline keyvault so that secrets
/ keys do not get stuck in this purgatory and require manual
portal-based intervention (typically requiring a full deletion of the
resourceGroup and all of its Resources because PurgeProtection cannot be
disabled on a Vault once its been created)