refactor: add a default $web container to bline
between the SPA and CDN components there is a chicken-and-egg problem
where each often wants the other to exist first.

To remove dependencies on eachother, this defines a $web container on
the baseline-provisioned storage account, which acts as a "default"
static web host in the event that another has not been specified.

It is not intended for use as a permanent static content host, but as
an intermediary location to allow for a CDN to exist (requires a
backend pool - i.e the $web container) prior to the configuration of
an SPA. Once configured, the content will be hosted in the SPA's
inidividual storage account.