refactor(cdn): handle updated spa storage accounts
In #248 the SPA component has been updated to configure unique storage
accounts per instance.

This updates the CDN component to use the new storage accounts as
backend pool configuration instead of the opsdata bucket.

In the event that a linked SPA is not found (so no backend pools)
rather than provide an error, the CDN will use a new default $web
container on the opsdata bucket as an intermediary step. This allows
the CDN to be deployed prior to the SPA, which is ideal given that
the CDN is a solution-level component and the SPA is application.